Add game logic and character management modules
- Introduced `game.py` for managing game state and player interactions.
- Added `day.py` to handle day phases, nominations, and voting mechanics.
- Created `traveler_vote.py` for managing traveler nominations and votes.
- Implemented `whisper_mode.py` to define whisper modes for player communication.
- Added utility functions in `game_utils.py` for updating game presence and managing backups.
- Created `script.py` to manage character scripts and their abilities.
- Added tests for new modules to ensure functionality and correctness.
