+++ bot_impl.py
@@ -6,945 +6,26 @@ import itertools
 import math
 import os
 from collections import OrderedDict
-from typing import Optional, Protocol, Sequence, TypeVar
+from typing import Optional, TypeVar, Protocol, Sequence
 
 import dill
 import discord
-from discord import Member, TextChannel, User
+from discord import User, Member, TextChannel
 
 import global_vars
 from bot_client import client, logger
 from config import *
 from model.channels import ChannelManager
-from model.characters import (
-    AbilityModifier, Amnesiac, Banshee, Character, DayEndModifier, DayStartModifier,
-    Demon, Minion, NominationModifier, NomsCalledModifier, Outsider,
-    SeatingOrderModifier, Storyteller, Townsfolk, Traveler, VoteBeginningModifier,
-    VoteModifier, Voudon, the_ability
-)
-from model.player import STORYTELLER_ALIGNMENT, Player
-from model.settings import GameSettings, GlobalSettings
+from model.characters import Character, AbilityModifier, Amnesiac, Minion, Demon, Outsider, Townsfolk, \
+    SeatingOrderModifier, Traveler, Storyteller, Banshee
+from model.game import Game, NULL_GAME, Script
+from model.game.vote import Vote, in_play_voudon, remove_banshee_nomination
+from model.player import Player, STORYTELLER_ALIGNMENT
+from model.settings import GlobalSettings, GameSettings
 from time_utils import parse_deadline
+from utils.character_utils import the_ability
 from utils.message_utils import safe_send
 from utils.player_utils import check_and_print_if_one_or_zero_to_check_in
-
-
-class WhisperMode:
-    ALL = 'all'
-    NEIGHBORS = 'neighbors'
-    STORYTELLERS = 'storytellers'
-
-
-### Classes
-class Game:
-    def __init__(self, seatingOrder, seatingOrderMessage, script, skip_storytellers=False):
-        self.days = []
-        self.isDay = False
-        self.script = script
-        self.seatingOrder = seatingOrder
-        self.whisper_mode = WhisperMode.ALL
-        self.seatingOrderMessage = seatingOrderMessage
-        self.storytellers = [
-            Player(Storyteller, STORYTELLER_ALIGNMENT, person, st_channel=None, position=None)
-            for person in global_vars.gamemaster_role.members
-        ] if not skip_storytellers else []
-        self.show_tally = False
-        self.has_automated_life_and_death = False
-
-    async def end(self, winner):
-        # Ends the game
-
-        # remove roles
-        for person in self.seatingOrder:
-            await person.wipe_roles()
-
-        # unpin messages
-        for msg in await global_vars.channel.pins():
-            if msg.created_at >= self.seatingOrderMessage.created_at:
-                await msg.unpin()
-
-        if global_vars.whisper_channel:
-            for msg in await global_vars.whisper_channel.pins():
-                await msg.unpin()
-
-        # announcement
-        winner = winner.lower()
-        await safe_send(
-            global_vars.channel,
-            f"{global_vars.player_role.mention}, {'The game is over.' if winner == 'tie' else f'{winner} has won.'} Good game!",
-        )
-
-        """
-        # save backup
-        i = 0
-        while True:
-            i += 1
-            if not os.path.isfile('game_{}.pckl'.format(str(i))):
-                break
-        backup('game_{}.pckl'.format(str(i)))
-        """
-
-        # delete old backup
-        remove_backup("current_game.pckl")
-
-        # turn off
-        global_vars.game = NULL_GAME
-        await update_presence(client)
-
-    async def reseat(self, newSeatingOrder):
-        # Reseats the table
-
-        # Seating order
-        self.seatingOrder = newSeatingOrder
-
-        # Seating order message
-        messageText = "**Seating Order:**"
-        for index, person in enumerate(self.seatingOrder):
-
-            if person.is_ghost:
-                if person.dead_votes <= 0:
-                    messageText += "\n{}".format("~~" + person.display_name + "~~ X")
-                else:
-                    messageText += "\n{}".format(
-                        "~~" + person.display_name + "~~ " + "O" * person.dead_votes
-                    )
-
-            else:
-                messageText += "\n{}".format(person.display_name)
-
-            if isinstance(person.character, SeatingOrderModifier):
-                messageText += person.character.seating_order_message(self.seatingOrder)
-
-            person.position = index
-
-        await self.seatingOrderMessage.edit(content=messageText)
-        #  fixme: reorder_channels is doing more work than needed. until that is fixed, don't reorder on reseat.
-        # await reorder_channels([x.st_channel for x in self.seatingOrder])
-
-    async def add_traveler(self, person):
-        self.seatingOrder.insert(person.position, person)
-        await person.user.add_roles(global_vars.player_role, global_vars.traveler_role)
-        await self.reseat(self.seatingOrder)
-        await safe_send(
-            global_vars.channel,
-            "{} has joined the town as the {}.".format(
-                person.display_name, person.character.role_name
-            ),
-        )
-
-    async def remove_traveler(self, person):
-        self.seatingOrder.remove(person)
-        await person.user.remove_roles(global_vars.player_role, global_vars.traveler_role)
-        await self.reseat(self.seatingOrder)
-        announcement = await safe_send(
-            global_vars.channel, "{} has left the town.".format(person.display_name)
-        )
-        await announcement.pin()
-
-    async def start_day(self, kills=None, origin=None):
-        if kills is None:
-            kills = []
-
-        for person in global_vars.game.seatingOrder:
-            await person.morning()
-            if isinstance(person.character, DayStartModifier):
-                if not await person.character.on_day_start(origin, kills):
-                    return
-
-        deaths = [await person.kill() for person in kills]
-        if deaths == [] and len(self.days) > 0:
-            no_kills = await safe_send(global_vars.channel, "No one has died.")
-            await no_kills.pin()
-        await safe_send(
-            global_vars.channel,
-            "{}, wake up! Message the storytellers to set default votes for today.".format(
-                global_vars.player_role.mention
-            ),
-        )
-        self.days.append(Day())
-        self.isDay = True
-
-        if global_vars.whisper_channel:
-            message = await safe_send(global_vars.whisper_channel, f"Start of day {len(self.days)}")
-            await message.pin()
-
-        await update_presence(client)
-
-
-class Script:
-    # Stores booleans for characters which modify the game rules from the script
-
-    def __init__(self, scriptList):
-        self.isAtheist = "atheist" in scriptList
-        self.isWitch = "witch" in scriptList
-        self.list = scriptList
-
-
-class Day:
-    # Stores information about a specific day
-
-    def __init__(self):
-        self.isExecutionToday = False
-        self.isNoms = False
-        self.isPms = True
-        self.votes = []
-        self.voteEndMessages = []
-        self.deadlineMessages = []
-        self.skipMessages = []
-        self.aboutToDie = None
-        self.riot_active = False
-        self.st_riot_kill_override = False
-
-    async def open_pms(self):
-        # Opens PMs
-        self.isPms = True
-        for memb in global_vars.gamemaster_role.members:
-            await safe_send(memb, "PMs are now open.")
-        await update_presence(client)
-
-    async def open_noms(self):
-        # Opens nominations
-        self.isNoms = True
-        if len(self.votes) == 0:
-            for person in global_vars.game.seatingOrder:
-                if isinstance(person.character, NomsCalledModifier):
-                    person.character.on_noms_called()
-        for memb in global_vars.gamemaster_role.members:
-            await safe_send(memb, "Nominations are now open.")
-        await update_presence(client)
-
-    async def close_pms(self):
-        # Closes PMs
-        self.isPms = False
-        for memb in global_vars.gamemaster_role.members:
-            await safe_send(memb, "PMs are now closed.")
-        await update_presence(client)
-
-    async def close_noms(self):
-        # Closes nominations
-        self.isNoms = False
-        for memb in global_vars.gamemaster_role.members:
-            await safe_send(memb, "Nominations are now closed.")
-        await update_presence(client)
-
-    async def nomination(self, nominee, nominator):
-        global_vars.game.whisper_mode = WhisperMode.NEIGHBORS
-        await self.close_noms()
-        # todo: if organ grinder ability is active, then this first message should not be output.
-        if not nominee:
-            self.votes.append(Vote(nominee, nominator))
-            if self.aboutToDie is not None:
-                announcement = await safe_send(
-                    global_vars.channel,
-                    "{}, the storytellers have been nominated by {}. {} to tie, {} to execute.".format(
-                        global_vars.player_role.mention,
-                        nominator.display_name if nominator else "the storytellers",
-                        str(
-                            int(
-                                math.ceil(
-                                    max(
-                                        self.aboutToDie[1].votes,
-                                        self.votes[-1].majority,
-                                    )
-                                )
-                            )
-                        ),
-                        str(int(math.ceil(self.aboutToDie[1].votes + 1))),
-                    ),
-                )
-            else:
-                announcement = await safe_send(
-                    global_vars.channel,
-                    "{}, the storytellers have been nominated by {}. {} to execute.".format(
-                        global_vars.player_role.mention,
-                        nominator.display_name if nominator else "the storytellers",
-                        str(int(math.ceil(self.votes[-1].majority))),
-                    ),
-                )
-            await announcement.pin()
-            if nominator and nominee and not isinstance(nominee.character, Traveler):
-                nominator.can_nominate = False
-            proceed = True
-            # FIXME:there might be a case where a player earlier in the seating order makes the nomination not proceed
-            #  but one later in the seating order may be relevant. Short circuit here stops two Riot messages, e.g.
-            #  There may need to be some rework based on NominationModifier priority
-            for person in global_vars.game.seatingOrder:
-                if isinstance(person.character, NominationModifier) and proceed:
-                    proceed = await person.character.on_nomination(
-                        nominee, nominator, proceed
-                    )
-            if not proceed:
-                # do not proceed with collecting votes
-                return
-        elif isinstance(nominee.character, Traveler):
-            nominee.can_be_nominated = False
-            self.votes.append(TravelerVote(nominee, nominator))
-            announcement = await safe_send(
-                global_vars.channel,
-                "{}, {} has called for {}'s exile. {} to exile.".format(
-                    global_vars.player_role.mention,
-                    nominator.display_name if nominator else "The storytellers",
-                    nominee.user.mention,
-                    str(int(math.ceil(self.votes[-1].majority))),
-                ),
-            )
-            await announcement.pin()
-        else:
-            nominee.can_be_nominated = False
-            self.votes.append(Vote(nominee, nominator))
-            # FIXME:there might be a case where a player earlier in the seating order makes the nomination not proceed
-            #  but one later in the seating order may be relevant. Short circuit here stops two Riot messages, e.g.
-            #  There may need to be some rework based on NominationModifier priority
-            proceed = True
-            for person in global_vars.game.seatingOrder:
-                if isinstance(person.character, NominationModifier) and proceed:
-                    proceed = await person.character.on_nomination(
-                        nominee, nominator, proceed
-                    )
-            if not proceed:
-                # do not proceed with collecting votes
-                return
-            if self.aboutToDie is not None:
-                announcement = await safe_send(
-                    global_vars.channel,
-                    "{}, {} has been nominated by {}. {} to tie, {} to execute.".format(
-                        global_vars.player_role.mention,
-                        nominee.user.mention
-                        if not nominee.user in global_vars.gamemaster_role.members
-                        else "the storytellers",
-                        nominator.display_name if nominator else "the storytellers",
-                        str(
-                            int(
-                                math.ceil(
-                                    max(
-                                        self.aboutToDie[1].votes,
-                                        self.votes[-1].majority,
-                                    )
-                                )
-                            )
-                        ),
-                        str(int(math.ceil(self.aboutToDie[1].votes + 1))),
-                    ),
-                )
-            else:
-                announcement = await safe_send(
-                    global_vars.channel,
-                    "{}, {} has been nominated by {}. {} to execute.".format(
-                        global_vars.player_role.mention,
-                        nominee.user.mention
-                        if not nominee.user in global_vars.gamemaster_role.members
-                        else "the storytellers",
-                        nominator.display_name if nominator else "the storytellers",
-                        str(int(math.ceil(self.votes[-1].majority))),
-                    ),
-                )
-            await announcement.pin()
-            if nominator:
-                nominator.can_nominate = False
-            proceed = True
-            # FIXME:there might be a case where a player earlier in the seating order makes the nomination not proceed
-            #  but one later in the seating order may be relevant. Short circuit here stops two Riot messages, e.g.
-            #  There may need to be some rework based on NominationModifier priority
-            for person in global_vars.game.seatingOrder:
-                if isinstance(person.character, NominationModifier) and proceed:
-                    proceed = await person.character.on_nomination(
-                        nominee, nominator, proceed
-                    )
-            if not proceed:
-                # do not proceed with collecting user input for this vote
-                return
-
-        if (global_vars.game.show_tally):
-            message_tally = {X: 0 for X in itertools.combinations(global_vars.game.seatingOrder, 2)}
-
-            has_had_multiple_votes = len(self.votes) > 1
-            last_vote_message = None if not has_had_multiple_votes else await global_vars.channel.fetch_message(self.votes[-2].announcements[0])
-
-            for person in global_vars.game.seatingOrder:
-                for msg in person.message_history:
-                    if msg["from_player"] == person:
-                        if has_had_multiple_votes:
-                            if msg["time"] >= last_vote_message.created_at:
-                                if (person, msg["to_player"]) in message_tally:
-                                    message_tally[(person, msg["to_player"])] += 1
-                                elif (msg["to_player"], person) in message_tally:
-                                    message_tally[(msg["to_player"], person)] += 1
-                                else:
-                                    message_tally[(person, msg["to_player"])] = 1
-                        else:
-                            if msg["day"] == len(global_vars.game.days):
-                                if (person, msg["to_player"]) in message_tally:
-                                    message_tally[(person, msg["to_player"])] += 1
-                                elif (msg["to_player"], person) in message_tally:
-                                    message_tally[(msg["to_player"], person)] += 1
-                                else:
-                                    message_tally[(person, msg["to_player"])] = 1
-            sorted_tally = sorted(message_tally.items(), key=lambda x: -x[1])
-            messageText = "**Message Tally:**"
-            for pair in sorted_tally:
-                if pair[1] > 0:
-                    messageText += "\n> {person1} - {person2}: {n}".format(
-                        person1=pair[0][0].display_name, person2=pair[0][1].display_name, n=pair[1]
-                    )
-                else:
-                    messageText += "\n> All other pairs: 0"
-                    break
-            await safe_send(global_vars.channel, messageText)
-
-        self.votes[-1].announcements.append(announcement.id)
-        await self.votes[-1].call_next()
-
-    async def end(self):
-        # Ends the day
-
-        for person in global_vars.game.seatingOrder:
-            if isinstance(person.character, DayEndModifier):
-                person.character.on_day_end()
-
-        for msg in self.voteEndMessages:
-            try:
-                await (await global_vars.channel.fetch_message(msg)).unpin()
-            except discord.errors.NotFound:
-                print("Missing message: ", str(msg))
-
-        for msg in self.deadlineMessages:
-            try:
-                await (await global_vars.channel.fetch_message(msg)).unpin()
-            except discord.errors.NotFound:
-                print("Missing message: ", str(msg))
-
-        for msg in self.skipMessages:
-            try:
-                await (await global_vars.channel.fetch_message(msg)).unpin()
-            except discord.errors.NotFound:
-                print("Missing message: ", str(msg))
-
-        global_vars.game.isDay = False
-        global_vars.game.whisper_mode = WhisperMode.ALL
-        self.isNoms = False
-        self.isPms = False
-
-        if not self.isExecutionToday:
-            await safe_send(global_vars.channel, "No one was executed.")
-
-        await safe_send(global_vars.channel, "{}, go to sleep!".format(global_vars.player_role.mention))
-
-        if not global_vars.game.days[-1].riot_active:
-            if (global_vars.game.show_tally):
-                message_tally = {X: 0 for X in itertools.combinations(global_vars.game.seatingOrder, 2)}
-                has_had_multiple_votes = len(self.votes) > 0
-
-                last_vote_message = None if not has_had_multiple_votes else (
-                    await global_vars.channel.fetch_message(self.votes[-1].announcements[0]) if self.votes and self.votes[-1].announcements else None
-                )
-                for person in global_vars.game.seatingOrder:
-                    for msg in person.message_history:
-                        if msg["from_player"] == person:
-                            if has_had_multiple_votes:
-                                if msg["time"] >= last_vote_message.created_at:
-                                    if (person, msg["to_player"]) in message_tally:
-                                        message_tally[(person, msg["to_player"])] += 1
-                                    elif (msg["to_player"], person) in message_tally:
-                                        message_tally[(msg["to_player"], person)] += 1
-                                    else:
-                                        message_tally[(person, msg["to_player"])] = 1
-                            else:
-                                if msg["day"] == len(global_vars.game.days):
-                                    if (person, msg["to_player"]) in message_tally:
-                                        message_tally[(person, msg["to_player"])] += 1
-                                    elif (msg["to_player"], person) in message_tally:
-                                        message_tally[(msg["to_player"], person)] += 1
-                                    else:
-                                        message_tally[(person, msg["to_player"])] = 1
-                sorted_tally = sorted(message_tally.items(), key=lambda x: -x[1])
-                messageText = "**Message Tally:**"
-                for pair in sorted_tally:
-                    if pair[1] > 0:
-                        messageText += "\n> {person1} - {person2}: {n}".format(
-                            person1=pair[0][0].display_name, person2=pair[0][1].display_name, n=pair[1]
-                        )
-                    else:
-                        messageText += "\n> All other pairs: 0"
-                        break
-                await safe_send(global_vars.channel, messageText)
-
-        await update_presence(client)
-
-
-class Vote:
-    # Stores information about a specific vote
-
-    def __init__(self, nominee, nominator):
-        self.nominee = nominee
-        self.nominator = nominator
-        if self.nominee is not None:
-            ordered_voters = (
-                    global_vars.game.seatingOrder[global_vars.game.seatingOrder.index(self.nominee) + 1:]
-                    + global_vars.game.seatingOrder[: global_vars.game.seatingOrder.index(self.nominee) + 1]
-            )
-        else:
-            ordered_voters = global_vars.game.seatingOrder
-        # augment order with Banshees
-        order_with_banshees = []
-        for person in ordered_voters:
-            order_with_banshees.append(person)
-            banshee_ability = the_ability(person.character, Banshee)
-            if banshee_ability and banshee_ability.is_screaming:
-                order_with_banshees.append(person)
-
-        self.order = order_with_banshees
-
-        self.votes = 0
-        self.voted = []
-        self.history = []
-        self.announcements = []
-        self.presetVotes = {}
-        self.values = {person: (0, 1) for person in self.order}
-        self.majority = 0.0
-        for person in self.order:
-            if not person.is_ghost:
-                self.majority += 0.5
-        for person in global_vars.game.seatingOrder:
-            if isinstance(person.character, VoteBeginningModifier):
-                (
-                    self.order,
-                    self.values,
-                    self.majority,
-                ) = person.character.modify_vote_values(
-                    self.order, self.values, self.majority
-                )
-        self.position = 0
-        self.done = False
-
-    async def call_next(self):
-        # Calls for person to vote
-
-        toCall = self.order[self.position]
-        player_banshee_ability = the_ability(toCall.character, Banshee)
-        player_is_active_banshee = player_banshee_ability and player_banshee_ability.is_screaming
-        voudon_is_active = in_play_voudon()
-        for person in global_vars.game.seatingOrder:
-            if isinstance(person.character, VoteModifier):
-                person.character.on_vote_call(toCall)
-        if toCall.is_ghost and toCall.dead_votes < 1 and not player_is_active_banshee and not voudon_is_active:
-            await self.vote(0)
-            return
-        if toCall.user.id in self.presetVotes:
-            preset_player_vote = self.presetVotes[toCall.user.id]
-            self.presetVotes[toCall.user.id] -= 1
-            await self.vote(int(preset_player_vote > 0))
-            return
-        await safe_send(
-            global_vars.channel,
-            "{}, your vote on {}.".format(
-                toCall.user.mention,
-                self.nominee.display_name if self.nominee else "the storytellers",
-            ),
-        )
-        global_settings: GlobalSettings = GlobalSettings.load()
-        default = global_settings.get_default_vote(toCall.user.id)
-        if default:
-            time = default[1]
-            await safe_send(toCall.user, "Will enter a {} vote in {} minutes.".format(
-                ["no", "yes"][default[0]], str(int(default[1] / 60))
-            ))
-            for memb in global_vars.gamemaster_role.members:
-                await safe_send(
-                    memb,
-                    "{}'s vote. Their default is {} in {} minutes.".format(
-                        toCall.display_name,
-                        ["no", "yes"][default[0]],
-                        str(int(default[1] / 60)),
-                    ),
-                )
-            await asyncio.sleep(time)
-            if toCall == global_vars.game.days[-1].votes[-1].order[global_vars.game.days[-1].votes[-1].position]:
-                await self.vote(default[0])
-        else:
-            for memb in global_vars.gamemaster_role.members:
-                await safe_send(
-                    memb, "{}'s vote. They have no default.".format(toCall.display_name)
-                )
-
-    async def vote(self, vt, operator=None):
-        # Executes a vote. vt is binary -- 0 if no, 1 if yes
-        # Voter
-        voter = self.order[self.position]
-
-        potential_banshee = the_ability(voter.character, Banshee)
-
-        # see if a Voudon is in play
-        voudon_in_play = in_play_voudon()
-        player_is_active_banshee = potential_banshee and voter.character.is_screaming
-        # Check dead votes
-        if vt == 1 and voter.is_ghost and voter.dead_votes < 1 and not (player_is_active_banshee and not potential_banshee.is_poisoned) and not voudon_in_play:
-            if not operator:
-                await safe_send(voter.user, "You do not have any dead votes. Entering a no vote.")
-                await self.vote(0)
-            else:
-                await safe_send(
-                    operator,
-                    "{} does not have any dead votes. They must vote no. If you want them to vote yes, add a dead vote first:\n```\n@givedeadvote [player]\n```".format(
-                        voter.display_name
-                    ),
-                )
-            return
-        if vt == 1 and voter.is_ghost and not (player_is_active_banshee and not potential_banshee.is_poisoned) and not voudon_in_play:
-            await voter.remove_dead_vote()
-
-        # On vote character powers
-        for person in global_vars.game.seatingOrder:
-            if isinstance(person.character, VoteModifier):
-                person.character.on_vote()
-
-        # Vote tracking
-        self.history.append(vt)
-        self.votes += self.values[voter][vt]
-        if vt == 1:
-            self.voted.append(voter)
-
-        # Announcement
-        text = "yes" if vt == 1 else "no"
-        self.announcements.append(
-            (
-                await safe_send(
-                    global_vars.channel,
-                    "{} votes {}. {} votes.".format(voter.display_name, text, str(self.votes)),
-                )
-            ).id
-        )
-        await (await global_vars.channel.fetch_message(self.announcements[-1])).pin()
-
-        # Next vote
-        self.position += 1
-        if self.position == len(self.order):
-            await self.end_vote()
-            return
-        await self.call_next()
-
-    async def end_vote(self):
-        # When the vote is over
-        tie = False
-        aboutToDie = global_vars.game.days[-1].aboutToDie
-        if self.votes >= self.majority:
-            if aboutToDie is None:
-                dies = True
-            elif self.votes > aboutToDie[1].votes:
-                dies = True
-            elif self.votes == aboutToDie[1].votes:
-                dies = False
-                tie = True
-            else:
-                dies = False
-        else:
-            dies = False
-        for person in global_vars.game.seatingOrder:
-            if isinstance(person.character, VoteModifier):
-                dies, tie = person.character.on_vote_conclusion(dies, tie)
-        for person in global_vars.game.seatingOrder:
-            person.riot_nominee = False
-        the_voters = self.voted
-        # remove duplicate voters
-        the_voters = list(OrderedDict.fromkeys(the_voters))
-        if len(the_voters) == 0:
-            text = "no one"
-        elif len(the_voters) == 1:
-            text = the_voters[0].display_name
-        elif len(the_voters) == 2:
-            text = the_voters[0].display_name + " and " + the_voters[1].display_name
-        else:
-            text = (", ".join([x.display_name for x in the_voters[:-1]]) + ", and " + the_voters[-1].display_name)
-        if dies:
-            if aboutToDie is not None and aboutToDie[0] is not None:
-                msg = await global_vars.channel.fetch_message(
-                    global_vars.game.days[-1].voteEndMessages[
-                        global_vars.game.days[-1].votes.index(aboutToDie[1])
-                    ]
-                )
-                await msg.edit(
-                    content=msg.content[:-31] + " They are not about to be executed."
-                )
-            global_vars.game.days[-1].aboutToDie = (self.nominee, self)
-            announcement = await safe_send(
-                global_vars.channel,
-                "{} votes on {} (nominated by {}): {}. They are about to be executed.".format(
-                    str(self.votes),
-                    self.nominee.display_name if self.nominee else "the storytellers",
-                    self.nominator.display_name if self.nominator else "the storytellers",
-                    text,
-                ),
-            )
-        elif tie:
-            if aboutToDie is not None:
-                msg = await global_vars.channel.fetch_message(
-                    global_vars.game.days[-1].voteEndMessages[
-                        global_vars.game.days[-1].votes.index(aboutToDie[1])
-                    ]
-                )
-                await msg.edit(
-                    content=msg.content[:-31] + " No one is about to be executed."
-                )
-            global_vars.game.days[-1].aboutToDie = (None, self)
-            announcement = await safe_send(
-                global_vars.channel,
-                "{} votes on {} (nominated by {}): {}. No one is about to be executed.".format(
-                    str(self.votes),
-                    self.nominee.display_name if self.nominee else "the storytellers",
-                    self.nominator.display_name if self.nominator else "the storytellers",
-                    text,
-                ),
-            )
-        else:
-            announcement = await safe_send(
-                global_vars.channel,
-                "{} votes on {} (nominated by {}): {}. They are not about to be executed.".format(
-                    str(self.votes),
-                    self.nominee.display_name if self.nominee else "the storytellers",
-                    self.nominator.display_name if self.nominator else "the storytellers",
-                    text,
-                ),
-            )
-
-        await announcement.pin()
-        global_vars.game.days[-1].voteEndMessages.append(announcement.id)
-
-        for msg in self.announcements:
-            try:
-                await (await global_vars.channel.fetch_message(msg)).unpin()
-            except discord.errors.NotFound:
-                print("Missing message: ", str(msg))
-
-        self.done = True
-
-        await global_vars.game.days[-1].open_noms()
-        await global_vars.game.days[-1].open_pms()
-
-    async def preset_vote(self, person, vt, operator=None):
-        # Check dead votes
-        banshee_ability = the_ability(person.character, Banshee)
-        banshee_override = banshee_ability and banshee_ability.is_screaming
-        if vt > 0 and person.is_ghost and person.dead_votes < 1 and not banshee_override:
-            if not operator:
-                await safe_send(person.user, "You do not have any dead votes. Please vote no.")
-            else:
-                await safe_send(
-                    operator,
-                    "{} does not have any dead votes. They must vote no.".format(
-                        person.display_name
-                    ),
-                )
-            return
-
-        self.presetVotes[person.user.id] = vt
-
-    async def cancel_preset(self, person):
-        if (person.user.id in self.presetVotes):
-            del self.presetVotes[person.user.id]
-
-    async def delete(self):
-        # Undoes an unintentional nomination
-
-        if self.nominator:
-            self.nominator.can_nominate = True
-        if self.nominee:
-            self.nominee.can_be_nominated = True
-
-        for msg in self.announcements:
-            try:
-                await (await global_vars.channel.fetch_message(msg)).unpin()
-            except discord.errors.NotFound:
-                pass
-
-        self.done = True
-
-        global_vars.game.days[-1].votes.remove(self)
-
-
-class TravelerVote:
-    # Stores information about a specific call for exile
-
-    def __init__(self, nominee, nominator):
-        self.nominee = nominee
-        self.nominator = nominator
-        self.order = (
-                global_vars.game.seatingOrder[global_vars.game.seatingOrder.index(self.nominee) + 1:]
-                + global_vars.game.seatingOrder[: global_vars.game.seatingOrder.index(self.nominee) + 1]
-        )
-        self.votes = 0
-        self.voted = []
-        self.history = []
-        self.announcements = []
-        self.presetVotes = {}
-        self.values = {person: (0, 1) for person in self.order}
-        self.majority = len(self.order) / 2
-        self.position = 0
-        self.done = False
-
-    async def call_next(self):
-        # Calls for person to vote
-
-        toCall = self.order[self.position]
-        if toCall.user.id in self.presetVotes:
-            await self.vote(self.presetVotes[toCall.user.id])
-            return
-        await safe_send(
-            global_vars.channel,
-            "{}, your vote on {}.".format(
-                toCall.user.mention,
-                self.nominee.display_name if self.nominee else "the storytellers",
-            ),
-        )
-        global_settings: GlobalSettings = GlobalSettings.load()
-        default = global_settings.get_default_vote(toCall.user.id)
-        if default:
-            time = default[1]
-            await safe_send(toCall.user, "Will enter a {} vote in {} minutes.".format(
-                ["no", "yes"][default[0]], str(int(default[1] / 60))
-            ))
-            await asyncio.sleep(time)
-            if toCall == global_vars.game.days[-1].votes[-1].order[global_vars.game.days[-1].votes[-1].position]:
-                await self.vote(default[0])
-            for memb in global_vars.gamemaster_role.members:
-                await safe_send(
-                    memb,
-                    "{}'s vote. Their default is {} in {} minutes.".format(
-                        toCall.display_name,
-                        ["no", "yes"][default[0]],
-                        str(int(default[1] / 60)),
-                    ),
-                )
-        else:
-            for memb in global_vars.gamemaster_role.members:
-                await safe_send(
-                    memb, "{}'s vote. They have no default.".format(toCall.display_name)
-                )
-
-    async def vote(self, vt, operator=None):
-        # Executes a vote. vt is binary -- 0 if no, 1 if yes
-
-        # Voter
-        voter = self.order[self.position]
-
-        # Vote tracking
-        self.history.append(vt)
-        self.votes += self.values[voter][vt]
-        if vt == 1:
-            self.voted.append(voter)
-
-        # Announcement
-        text = "yes" if vt == 1 else "no"
-        self.announcements.append(
-            (
-                await safe_send(
-                    global_vars.channel,
-                    "{} votes {}. {} votes.".format(voter.display_name, text, str(self.votes)),
-                )
-            ).id
-        )
-        await (await global_vars.channel.fetch_message(self.announcements[-1])).pin()
-
-        # Next vote
-        self.position += 1
-        if self.position == len(self.order):
-            await self.end_vote()
-            return
-        await self.call_next()
-
-    async def end_vote(self):
-        # When the vote is over
-        if len(self.voted) == 0:
-            text = "no one"
-        elif len(self.voted) == 1:
-            text = self.voted[0].display_name
-        elif len(self.voted) == 2:
-            text = self.voted[0].display_name + " and " + self.voted[1].display_name
-        else:
-            text = (", ".join([x.display_name for x in self.voted[:-1]]) + ", and " + self.voted[-1].display_name)
-        if self.votes >= self.majority:
-            announcement = await safe_send(
-                global_vars.channel,
-                "{} votes on {} (nominated by {}): {}.".format(
-                    str(self.votes),
-                    self.nominee.display_name if self.nominee else "the storytellers",
-                    self.nominator.display_name if self.nominator else "the storytellers",
-                    text,
-                ),
-            )
-        else:
-            announcement = await safe_send(
-                global_vars.channel,
-                "{} votes on {} (nominated by {}): {}. They are not exiled.".format(
-                    str(self.votes),
-                    self.nominee.display_name if self.nominee else "the storytellers",
-                    self.nominator.display_name if self.nominator else "the storytellers",
-                    text,
-                ),
-            )
-
-        await announcement.pin()
-        global_vars.game.days[-1].voteEndMessages.append(announcement.id)
-
-        for msg in self.announcements:
-            try:
-                await (await global_vars.channel.fetch_message(msg)).unpin()
-            except discord.errors.NotFound:
-                print("Missing message: ", str(msg))
-
-        self.done = True
-
-        await global_vars.game.days[-1].open_noms()
-        await global_vars.game.days[-1].open_pms()
-
-    async def preset_vote(self, person, vt, operator=None):
-        self.presetVotes[person.user.id] = vt
-
-    async def cancel_preset(self, person):
-        del self.presetVotes[person.user.id]
-
-    async def delete(self):
-        # Undoes an unintentional nomination
-
-        if self.nominator:
-            self.nominator.can_nominate = True
-        if self.nominee:
-            self.nominee.can_be_nominated = True
-
-        for msg in self.announcements:
-            try:
-                await (await global_vars.channel.fetch_message(msg)).unpin()
-            except discord.errors.NotFound:
-                print("Missing message: ", str(msg))
-
-        self.done = True
-
-        global_vars.game.days[-1].votes.remove(self)
-
-
-# Player class has been extracted to model/player.py
-
-# Character classes have been moved to model/characters/ directory
-
-# has_ability and the_ability functions have been moved to model/characters/base.py
-
-# Character Helper functions
-
-def in_play_voudon():
-    """Check if a Voudon is in play.
-
-    Returns:
-        The Voudon player if one is in play and not a ghost, otherwise None
-    """
-    return next((x for x in global_vars.game.seatingOrder if the_ability(x.character, Voudon) and not x.is_ghost), None)
-
-
-def remove_banshee_nomination(banshee_ability_of_player):
-    """Remove a nomination from a Banshee.
-
-    Args:
-        banshee_ability_of_player: The Banshee ability
-    """
-    if banshee_ability_of_player and banshee_ability_of_player.is_screaming:
-        banshee_ability_of_player.remaining_nominations -= 1
 
 ### API Stuff
 try:
@@ -1172,31 +253,13 @@ async def cannot_nominate(user):
 
 async def update_presence(client):
     # Updates Discord Presence
-
-    if global_vars.game is NULL_GAME:
-        await client.change_presence(
-            status=discord.Status.dnd, activity=discord.Game(name="No ongoing game!")
-        )
-    elif global_vars.game.isDay == False:
-        await client.change_presence(
-            status=discord.Status.idle, activity=discord.Game(name="It's nighttime!")
-        )
-    else:
-        clopen = ["Closed", "Open"]
-
-        whisper_state = "to " + global_vars.game.whisper_mode if global_vars.game.days[-1].isPms and global_vars.game.whisper_mode != WhisperMode.ALL else clopen[
-            global_vars.game.days[-1].isPms]
-        status = "PMs {}, Nominations {}!".format(whisper_state, clopen[global_vars.game.days[-1].isNoms])
-        await client.change_presence(
-            status=discord.Status.online,
-            activity=discord.Game(
-                name=status
-            ),
-        )
+    from utils.game_utils import update_presence as utils_update_presence
+    await utils_update_presence(client)
 
 
 def backup(fileName):
     # Backs up the game-state
+    from model.game.game import NULL_GAME
     if not global_vars.game or global_vars.game is NULL_GAME:
         return
 
@@ -1218,6 +281,8 @@ def backup(fileName):
 
 async def load(fileName):
     # Loads the game-state
+    from model.game.game import Game
+    from model.game.script import Script
 
     with open(fileName, "rb") as file:
         objects = dill.load(file)
@@ -1239,13 +304,8 @@ async def load(fileName):
 
 
 def remove_backup(fileName):
-    os.remove(fileName)
-    for obj in [
-        x
-        for x in dir(global_vars.game)
-        if not x.startswith("__") and not callable(getattr(global_vars.game, x))
-    ]:
-        os.remove(obj + "_" + fileName)
+    from utils.game_utils import remove_backup as utils_remove_backup
+    utils_remove_backup(fileName)
 
 
 def find_all(p, s):
@@ -1284,12 +344,14 @@ async def safe_send(target: discord.abc.Messageable, msg: str):
         else:
             raise (e)
 
-NULL_GAME = Game(seatingOrder=[], seatingOrderMessage=0, script=[], skip_storytellers=True)
+
+# NULL_GAME is imported from model.game
 
 ### Event Handling
 @client.event
 async def on_ready():
     # On startup
+    from model.game.game import NULL_GAME
 
     global_vars.game = NULL_GAME
     global_vars.observer_role = None
@@ -1610,14 +672,14 @@ async def on_message(message):
                     )
 
                 game_settings = GameSettings.load()
-                st_channel = client.get_channel(game_settings.get_st_channel(player.user.id))
+                st_channel = client.get_channel(game_settings.get_st_channel(player.id))
                 if not st_channel:
                     st_channel = await ChannelManager(client).create_channel(game_settings, player)
                     await safe_send(message.author,
                                     f'Successfully created the channel https://discord.com/channels/{global_vars.server.id}/{st_channel.id}!')
 
                 await safe_send(
-                    player.user,
+                    player,
                     "Hello, {player_nick}! {storyteller_nick} welcomes you to Blood on the Clocktower on Discord! I'm {bot_nick}, the bot used on #{channel_name} in {server_name} to run games. Your Storyteller channel for this game is #{st_channel}\n\nThis is where you'll perform your private messaging during the game. To send a pm to a player, type `@pm [name]`.\n\nFor more info, type `@help`, or ask the storyteller(s): {storytellers}.".format(
                         bot_nick=bot_nick,
                         channel_name=channel_name,
@@ -2800,10 +1862,11 @@ async def on_message(message):
                     return
 
                 # argument is true or false
-                global_vars.game.script.isAtheist = argument.lower() == "true" or argument.lower() == "t"
+                global_vars.game.script.is_atheist = argument.lower() == "true" or argument.lower() == "t"
                 #  message storytellers that atheist game is set to false
                 for memb in global_vars.gamemaster_role.members:
-                    await safe_send(memb, "Atheist game is set to {} by {}".format(global_vars.game.script.isAtheist, message.author.display_name))
+                    await safe_send(memb, "Atheist game is set to {} by {}".format(global_vars.game.script.is_atheist,
+                                                                                   message.author.display_name))
                 pass
             elif command == "automatekills":
                 if global_vars.game is NULL_GAME:
@@ -3071,7 +2134,7 @@ async def on_message(message):
                         await safe_send(message.author, "You have already nominated.")
                         return
 
-                if global_vars.game.script.isAtheist:
+                if global_vars.game.script.is_atheist:
                     if story_teller_is_nominated:
                         if None in [x.nominee for x in global_vars.game.days[-1].votes]:
                             await safe_send(message.author, "The storytellers have already been nominated today.")
@@ -4166,52 +3229,9 @@ async def check_and_print_if_one_or_zero_to_check_in():
             await safe_send(memb, "Everyone has checked in!")
 
 
-def to_whisper_mode(argument):
-    """Convert a string argument to a WhisperMode enum value.
-    
-    Args:
-        argument: String input representing a whisper mode
-        
-    Returns:
-        WhisperMode enum value or None if not valid
-    """
-    if argument is None:
-        return None
+from model.game.whisper_mode import to_whisper_mode, chose_whisper_candidates
 
-    new_mode = None
-    try:
-        if WhisperMode.ALL.casefold() == argument.casefold():
-            new_mode = WhisperMode.ALL
-        elif WhisperMode.NEIGHBORS.casefold() == argument.casefold():
-            new_mode = WhisperMode.NEIGHBORS
-        elif WhisperMode.STORYTELLERS.casefold() == argument.casefold():
-            new_mode = WhisperMode.STORYTELLERS
-    except AttributeError:
-        # Handle cases where argument doesn't have casefold()
-        return None
-        
-    return new_mode
-
-
-async def chose_whisper_candidates(game, author):
-    if game.whisper_mode == WhisperMode.ALL:
-        return game.seatingOrder + game.storytellers
-    if game.whisper_mode == WhisperMode.STORYTELLERS:
-        return game.storytellers
-    if game.whisper_mode == WhisperMode.NEIGHBORS:
-        # determine neighbors
-        player_self = await get_player(author)
-        author_index = game.seatingOrder.index(player_self)
-        neighbor_left = game.seatingOrder[(author_index - 1) % len(game.seatingOrder)]
-        neighbor_right = game.seatingOrder[(author_index + 1) % len(game.seatingOrder)]
-        return [neighbor_left, player_self, neighbor_right] + game.storytellers
-
-
-async def is_storyteller(arg):
-    if arg in ["storytellers", "the storytellers", "storyteller", "the storyteller"]:
-        return True
-    options = await generate_possibilities(arg, global_vars.server.members)
-    return len(options) == 1 and global_vars.gamemaster_role in global_vars.server.get_member((options)[0].id).roles
+from model.game.vote import is_storyteller
 
 
 @client.event
@@ -4274,7 +3294,7 @@ async def on_message_edit(before, after):
                 await after.unpin()
                 return
 
-            if global_vars.game.script.isAtheist:
+            if global_vars.game.script.is_atheist:
                 storyteller_nomination = await is_storyteller(argument)
                 if storyteller_nomination:
                     if None in [x.nominee for x in global_vars.game.days[-1].votes]:
