+++ model/game/__init__.py
@@ -1,0 +1,17 @@
+# Import modules directly
+from model.game.day import Day
+# Import classes from modules
+from model.game.game import Game, NULL_GAME
+from model.game.script import Script
+from model.game.traveler_vote import TravelerVote
+from model.game.vote import Vote
+from model.game.whisper_mode import WhisperMode
+
+__all__ = [
+    'WhisperMode',
+    'Game',
+    'NULL_GAME',
+    'Script',
+    'Day',
+    'Vote',
+]
